Unite FireAndForgetJobProcessor and DelayedJobProcessor
